Går inte längre sönder ifall localStorage är Null
