+++ car-pool/src/App.js
@@ -5,6 +5,7 @@ import BookTravel from './Pages/BookTravel.js';
 import BookPickup from './Pages/BookPickup.js';
 import Layout from './Pages/Layout.js';
 import ListBookPickup from "./Pages/ListBookPickup"
+import ListBookTravel from "./Pages/ListBookTravel"
 
 function App() {
   return (
@@ -17,7 +18,8 @@ function App() {
             <Route index element={<div>home</div>} />
             <Route path='/1' element={<BookTravel />} />
             <Route path='/2' element={<BookPickup />} />
-            <Route path='/3' element={<ListBookPickup/>} />
+            <Route path='/3' element={<ListBookTravel/>} />
+            <Route path='/4' element={<ListBookPickup/>} />
           </Route>
         </Routes>
       </Router>
