+++ car-pool/src/Components/Menu.js
@@ -21,6 +21,9 @@ function Menu() {
         <button>Boka upphämtning</button>
       </NavLink>
       <NavLink to='/3'>
+        <button>Bokade resor</button>
+      </NavLink>
+      <NavLink to='/4'>
         <button>Bokade upphämtningar</button>
       </NavLink>
     </div>
