+++ car-pool/src/Pages/ListBookTravel.js
@@ -1,0 +1,9 @@
+import { json } from "react-router-dom"
+import List from "../Components/List"
+
+function ListBookTravel() 
+{
+    return <List storageItem={"travel"}/>
+}
+
+export default ListBookTravel
